Move the test back, since it doesn't work in 1.2.1 mode
This is a test for unnamed kernels, so if we don't have unnamed kernels,
it isn't meaningful.
